Move alert templates to alerts service
