+++ server/py/services/alerts/crud/__init__.py
@@ -11,5 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .alert_template import AlertTemplates
 from .alerts import Alerts
 from .events import Events
